feat: refact kube forward code and add button to inspect logs
