+++ crates/kftray-tauri/src/kubeforward/commands.rs
@@ -1,0 +1,782 @@
+use std::collections::HashMap;
+use std::path::PathBuf;
+use std::sync::{
+    Arc,
+    Mutex,
+};
+use std::{
+    fs::File,
+    io::Read,
+    time::{
+        SystemTime,
+        UNIX_EPOCH,
+    },
+};
+
+use hostsfile::HostsBuilder;
+use kube::{
+    api::{
+        Api,
+        DeleteParams,
+        ListParams,
+    },
+    Client,
+};
+use kube_runtime::wait::conditions;
+use lazy_static::lazy_static;
+use rand::{
+    distributions::Alphanumeric,
+    Rng,
+};
+use tokio::task::JoinHandle;
+
+use crate::kubeforward::kubecontext::create_client_with_specific_context;
+use crate::{
+    config,
+    kubeforward::vx::Pod,
+    models::{
+        config::Config,
+        kube::{
+            Port,
+            PortForward,
+            Target,
+            TargetSelector,
+        },
+        response::CustomResponse,
+    },
+};
+
+lazy_static! {
+    pub static ref CHILD_PROCESSES: Arc<Mutex<HashMap<String, JoinHandle<()>>>> =
+        Arc::new(Mutex::new(HashMap::new()));
+}
+
+#[tauri::command]
+pub async fn start_port_udp_forward(configs: Vec<Config>) -> Result<Vec<CustomResponse>, String> {
+    let mut responses = Vec::new();
+    let mut errors = Vec::new();
+    let mut child_handles = Vec::new();
+
+    for config in configs.iter() {
+        let selector = TargetSelector::ServiceName(config.service.clone().unwrap());
+        let remote_port = Port::from(config.remote_port as i32);
+        let context_name = Some(config.context.clone());
+        let kubeconfig = Some(config.kubeconfig.clone());
+        let namespace = config.namespace.clone();
+        let target = Target::new(selector, remote_port, namespace);
+
+        log::info!("Remote Port: {}", config.remote_port);
+        log::info!("Local Port: {}", config.local_port);
+        log::debug!("Attempting to forward to service: {:?}", &config.service);
+
+        let local_address_clone = config.local_address.clone();
+
+        match PortForward::new(
+            target,
+            config.local_port,
+            local_address_clone,
+            context_name,
+            kubeconfig.flatten(),
+            config.id.unwrap_or_default(),
+        )
+        .await
+        {
+            Ok(port_forward) => match port_forward.port_forward_udp().await {
+                Ok((actual_local_port, handle)) => {
+                    log::info!(
+                        "UDP port forwarding is set up on local port: {:?} for service: {:?}",
+                        actual_local_port,
+                        &config.service
+                    );
+
+                    let handle_key = format!(
+                        "{}_{}",
+                        config.id.unwrap(),
+                        config.service.clone().unwrap_or_default()
+                    );
+                    CHILD_PROCESSES
+                        .lock()
+                        .unwrap()
+                        .insert(handle_key.clone(), handle);
+                    child_handles.push(handle_key.clone()); // Clone handle_key for tracking
+
+                    if config.domain_enabled.unwrap_or_default() {
+                        let hostfile_comment = format!(
+                            "kftray custom host for {} - {}",
+                            config.service.clone().unwrap_or_default(),
+                            config.id.unwrap_or_default()
+                        );
+
+                        let mut hosts_builder = HostsBuilder::new(hostfile_comment);
+
+                        if let Some(service_name) = &config.service {
+                            if let Some(local_address) = &config.local_address {
+                                match local_address.parse::<std::net::IpAddr>() {
+                                    Ok(ip_addr) => {
+                                        hosts_builder.add_hostname(
+                                            ip_addr,
+                                            config.alias.clone().unwrap_or_default(),
+                                        );
+                                        if let Err(e) = hosts_builder.write() {
+                                            let error_message = format!(
+                                                "Failed to write to the hostfile for {}: {}",
+                                                service_name, e
+                                            );
+                                            log::error!("{}", &error_message);
+                                            errors.push(error_message);
+
+                                            // Abort the child process due to critical error
+                                            if let Some(handle) =
+                                                CHILD_PROCESSES.lock().unwrap().remove(&handle_key)
+                                            {
+                                                handle.abort();
+                                            }
+                                            continue;
+                                        }
+                                    }
+                                    Err(_) => {
+                                        let warning_message =
+                                            format!("Invalid IP address format: {}", local_address);
+                                        log::warn!("{}", &warning_message);
+                                        errors.push(warning_message);
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    responses.push(CustomResponse {
+                        id: config.id,
+                        service: config.service.clone().unwrap(),
+                        namespace: config.namespace.clone(),
+                        local_port: actual_local_port,
+                        remote_port: config.remote_port,
+                        context: config.context.clone(),
+                        protocol: config.protocol.clone(),
+                        stdout: format!(
+                            "UDP forwarding from 127.0.0.1:{} -> {}:{}",
+                            actual_local_port,
+                            config.remote_port,
+                            config.service.clone().unwrap()
+                        ),
+                        stderr: String::new(),
+                        status: 0,
+                    });
+                }
+                Err(e) => {
+                    let error_message = format!(
+                        "Failed to start UDP port forwarding for service {}: {}",
+                        config.service.clone().unwrap_or_default(),
+                        e
+                    );
+                    log::error!("{}", &error_message);
+                    errors.push(error_message);
+                }
+            },
+            Err(e) => {
+                let error_message = format!(
+                    "Failed to create PortForward for service {}: {}",
+                    config.service.clone().unwrap_or_default(),
+                    e
+                );
+                log::error!("{}", &error_message);
+                errors.push(error_message);
+            }
+        }
+    }
+
+    if !errors.is_empty() {
+        // Abort all child processes if any critical error occurred
+        for handle_key in child_handles {
+            if let Some(handle) = CHILD_PROCESSES.lock().unwrap().remove(&handle_key) {
+                handle.abort();
+            }
+        }
+        return Err(errors.join("\n"));
+    }
+
+    if !responses.is_empty() {
+        log::info!("UDP port forwarding responses generated successfully.");
+    }
+
+    Ok(responses)
+}
+
+#[tauri::command]
+pub async fn start_port_forward(configs: Vec<Config>) -> Result<Vec<CustomResponse>, String> {
+    let mut responses = Vec::new();
+    let mut errors = Vec::new();
+    let mut child_handles = Vec::new();
+
+    for config in configs.iter() {
+        let selector = TargetSelector::ServiceName(config.service.clone().unwrap());
+        let remote_port = Port::from(config.remote_port as i32);
+        let context_name = Some(config.context.clone());
+        let kubeconfig = Some(config.kubeconfig.clone());
+        let namespace = config.namespace.clone();
+        let target = Target::new(selector, remote_port, namespace);
+
+        log::info!("Remote Port: {}", config.remote_port);
+        log::info!("Local Port: {}", config.local_port);
+        log::info!("Local Address: {:?}", config.local_address);
+
+        match PortForward::new(
+            target,
+            config.local_port,
+            config.local_address.clone(),
+            context_name.clone(),
+            kubeconfig.clone().flatten(),
+            config.id.unwrap_or_default(),
+        )
+        .await
+        {
+            Ok(port_forward) => match port_forward.port_forward().await {
+                Ok((actual_local_port, handle)) => {
+                    log::info!(
+                        "Port forwarding is set up on local port: {:?} for service: {:?}",
+                        actual_local_port,
+                        &config.service
+                    );
+
+                    let handle_key = format!(
+                        "{}_{}",
+                        config.id.unwrap(),
+                        config.service.clone().unwrap_or_default()
+                    );
+                    CHILD_PROCESSES
+                        .lock()
+                        .unwrap()
+                        .insert(handle_key.clone(), handle);
+                    child_handles.push(handle_key.clone());
+
+                    if config.domain_enabled.unwrap_or_default() {
+                        let hostfile_comment = format!(
+                            "kftray custom host for {} - {}",
+                            config.service.clone().unwrap_or_default(),
+                            config.id.unwrap_or_default()
+                        );
+
+                        let mut hosts_builder = HostsBuilder::new(hostfile_comment);
+
+                        if let Some(service_name) = &config.service {
+                            if let Some(local_address) = &config.local_address {
+                                match local_address.parse::<std::net::IpAddr>() {
+                                    Ok(ip_addr) => {
+                                        hosts_builder.add_hostname(
+                                            ip_addr,
+                                            config.alias.clone().unwrap_or_default(),
+                                        );
+                                        if let Err(e) = hosts_builder.write() {
+                                            let error_message = format!(
+                                                "Failed to write to the hostfile for {}: {}",
+                                                service_name, e
+                                            );
+                                            log::error!("{}", &error_message);
+                                            errors.push(error_message);
+
+                                            if let Some(handle) =
+                                                CHILD_PROCESSES.lock().unwrap().remove(&handle_key)
+                                            {
+                                                handle.abort();
+                                            }
+                                            continue;
+                                        }
+                                    }
+                                    Err(_) => {
+                                        let warning_message =
+                                            format!("Invalid IP address format: {}", local_address);
+                                        log::warn!("{}", &warning_message);
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    responses.push(CustomResponse {
+                        id: config.id,
+                        service: config.service.clone().unwrap(),
+                        namespace: config.namespace.clone(),
+                        local_port: actual_local_port,
+                        remote_port: config.remote_port,
+                        context: config.context.clone(),
+                        protocol: config.protocol.clone(),
+                        stdout: format!(
+                            "Forwarding from 127.0.0.1:{} -> {}:{}",
+                            actual_local_port,
+                            config.remote_port,
+                            config.service.clone().unwrap()
+                        ),
+                        stderr: String::new(),
+                        status: 0,
+                    });
+                }
+                Err(e) => {
+                    let error_message = format!("Failed to start port forwarding: {}", e);
+                    log::error!("{}", &error_message);
+                    errors.push(error_message);
+                }
+            },
+            Err(e) => {
+                let error_message = format!("Failed to create PortForward for service: {}", e);
+                log::error!("{}", &error_message);
+                errors.push(error_message);
+            }
+        }
+    }
+
+    if !errors.is_empty() {
+        for handle_key in child_handles {
+            if let Some(handle) = CHILD_PROCESSES.lock().unwrap().remove(&handle_key) {
+                handle.abort();
+            }
+        }
+        return Err(errors.join("\n"));
+    }
+
+    if !responses.is_empty() {
+        log::info!("Port forwarding responses generated successfully.");
+    }
+
+    Ok(responses)
+}
+
+#[tauri::command]
+pub async fn stop_all_port_forward() -> Result<Vec<CustomResponse>, String> {
+    log::info!("Attempting to stop all port forwards");
+
+    let mut responses = Vec::new();
+
+    let client = Client::try_default().await.map_err(|e| {
+        log::error!("Failed to create Kubernetes client: {}", e);
+        e.to_string()
+    })?;
+
+    // Drain the global hashmap to take ownership of all tasks
+    let handle_map: HashMap<String, tokio::task::JoinHandle<()>> =
+        CHILD_PROCESSES.lock().unwrap().drain().collect();
+
+    let configs_result = config::get_configs().await;
+    if let Err(e) = configs_result {
+        return Err(format!("Failed to retrieve configs: {}", e));
+    }
+    let configs = configs_result.unwrap();
+
+    for (composite_key, handle) in handle_map.iter() {
+        let ids: Vec<&str> = composite_key.split('_').collect();
+        if ids.len() != 2 {
+            log::error!(
+                "Invalid composite key format encountered: {}",
+                composite_key
+            );
+            continue;
+        }
+
+        let config_id_str = ids[0];
+        let service_id = ids[1];
+        let config_id_parsed = config_id_str.parse::<i64>().unwrap_or_default();
+
+        let config = configs
+            .iter()
+            .find(|c| c.id.map_or(false, |id| id == config_id_parsed));
+        if let Some(config) = config {
+            if config.domain_enabled.unwrap_or_default() {
+                let hostfile_comment =
+                    format!("kftray custom host for {} - {}", service_id, config_id_str);
+                let hosts_builder = HostsBuilder::new(&hostfile_comment);
+
+                if let Err(e) = hosts_builder.write() {
+                    log::error!("Failed to write to the hostfile for {}: {}", service_id, e);
+                    responses.push(CustomResponse {
+                        id: Some(config_id_parsed),
+                        service: service_id.to_string(),
+                        namespace: String::new(),
+                        local_port: 0,
+                        remote_port: 0,
+                        context: String::new(),
+                        protocol: String::new(),
+                        stdout: String::new(),
+                        stderr: e.to_string(),
+                        status: 1,
+                    });
+                    continue;
+                }
+            }
+        } else {
+            log::warn!("Config with id '{}' not found.", config_id_str);
+        }
+
+        log::info!(
+            "Aborting port forwarding task for config_id: {}",
+            config_id_str
+        );
+        handle.abort();
+
+        // Delete pods
+        let pods: Api<Pod> = Api::all(client.clone());
+        let lp = ListParams::default().labels(&format!("config_id={}", config_id_str));
+        log::info!(
+            "Listing pods with label selector: config_id={}",
+            config_id_str
+        );
+        let pod_list = pods.list(&lp).await.map_err(|e| {
+            log::error!("Error listing pods for config_id {}: {}", config_id_str, e);
+            e.to_string()
+        })?;
+
+        let username = whoami::username();
+        let pod_prefix = format!("kftray-forward-{}", username);
+
+        for pod in pod_list.items.into_iter() {
+            if let Some(pod_name) = pod.metadata.name {
+                log::info!("Found pod: {}", pod_name);
+                if pod_name.starts_with(&pod_prefix) {
+                    log::info!("Deleting pod: {}", pod_name);
+                    let namespace = pod
+                        .metadata
+                        .namespace
+                        .clone()
+                        .unwrap_or_else(|| "default".to_string());
+                    let pods_in_namespace: Api<Pod> = Api::namespaced(client.clone(), &namespace);
+                    let dp = DeleteParams {
+                        grace_period_seconds: Some(0),
+                        ..DeleteParams::default()
+                    };
+                    if let Err(e) = pods_in_namespace.delete(&pod_name, &dp).await {
+                        log::error!(
+                            "Failed to delete pod {} in namespace {}: {}",
+                            pod_name,
+                            namespace,
+                            e
+                        );
+                    } else {
+                        log::info!("Successfully deleted pod: {}", pod_name);
+                    }
+                }
+            }
+        }
+
+        responses.push(CustomResponse {
+            id: Some(config_id_parsed),
+            service: service_id.to_string(),
+            namespace: String::new(),
+            local_port: 0,
+            remote_port: 0,
+            context: String::new(),
+            protocol: String::new(),
+            stdout: String::from("Service port forwarding has been stopped"),
+            stderr: String::new(),
+            status: 0,
+        });
+    }
+
+    log::info!(
+        "Port forward stopping process completed with {} responses",
+        responses.len()
+    );
+
+    Ok(responses)
+}
+
+#[tauri::command]
+pub async fn stop_port_forward(
+    _service_name: String, config_id: String,
+) -> Result<CustomResponse, String> {
+    let composite_key = {
+        let child_processes = CHILD_PROCESSES.lock().unwrap();
+        child_processes
+            .keys()
+            .find(|key| key.starts_with(&format!("{}_", config_id)))
+            .map(|key| key.to_string())
+    };
+
+    if let Some(composite_key) = composite_key {
+        let handle = {
+            let mut child_processes = CHILD_PROCESSES.lock().unwrap();
+            child_processes.remove(&composite_key)
+        };
+
+        if let Some(handle) = handle {
+            handle.abort();
+
+            let (config_id_str, service_name) = composite_key.split_once('_').unwrap_or(("", ""));
+            let config_id_parsed = config_id_str.parse::<i64>().unwrap_or_default();
+
+            let configs_result = config::get_configs().await;
+            match configs_result {
+                Ok(configs) => {
+                    let config = configs
+                        .iter()
+                        .find(|c| c.id.map_or(false, |id| id == config_id_parsed));
+
+                    if let Some(config) = config {
+                        if config.domain_enabled.unwrap_or_default() {
+                            let hostfile_comment = format!(
+                                "kftray custom host for {} - {}",
+                                service_name, config_id_str
+                            );
+
+                            let hosts_builder = HostsBuilder::new(hostfile_comment);
+
+                            hosts_builder.write().map_err(|e| {
+                                log::error!(
+                                    "Failed to remove from the hostfile for {}: {}",
+                                    service_name,
+                                    e
+                                );
+
+                                e.to_string()
+                            })?;
+                        }
+                    } else {
+                        log::warn!("Config with id '{}' not found.", config_id_str);
+                    }
+
+                    Ok(CustomResponse {
+                        id: None,
+                        service: service_name.to_string(),
+                        namespace: String::new(),
+                        local_port: 0,
+                        remote_port: 0,
+                        context: String::new(),
+                        protocol: String::new(),
+                        stdout: String::from("Service port forwarding has been stopped"),
+                        stderr: String::new(),
+                        status: 0,
+                    })
+                }
+                Err(e) => Err(format!("Failed to retrieve configs: {}", e)),
+            }
+        } else {
+            Err(format!(
+                "Failed to stop port forwarding process for config_id '{}'",
+                config_id
+            ))
+        }
+    } else {
+        Err(format!(
+            "No port forwarding process found for config_id '{}'",
+            config_id
+        ))
+    }
+}
+
+fn get_pod_manifest_path() -> PathBuf {
+    let home_dir = dirs::home_dir().expect("Failed to resolve home directory");
+
+    home_dir.join(".kftray/proxy_manifest.json")
+}
+
+fn render_json_template(template: &str, values: &HashMap<&str, String>) -> String {
+    let mut rendered_template = template.to_string();
+
+    for (key, value) in values.iter() {
+        rendered_template = rendered_template.replace(&format!("{{{}}}", key), value);
+    }
+
+    rendered_template
+}
+
+#[tauri::command]
+
+pub async fn deploy_and_forward_pod(configs: Vec<Config>) -> Result<Vec<CustomResponse>, String> {
+    let mut responses: Vec<CustomResponse> = Vec::new();
+
+    for mut config in configs.into_iter() {
+        let client = if !config.context.is_empty() {
+            let kubeconfig = config.kubeconfig.clone();
+
+            create_client_with_specific_context(kubeconfig, &config.context)
+                .await
+                .map_err(|e| e.to_string())?
+        } else {
+            Client::try_default().await.map_err(|e| e.to_string())?
+        };
+
+        let timestamp = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .map_err(|e| e.to_string())?
+            .as_secs();
+
+        let random_string: String = rand::thread_rng()
+            .sample_iter(&Alphanumeric)
+            .take(6)
+            .map(char::from)
+            .map(|c| c.to_ascii_lowercase())
+            .collect();
+
+        let username = whoami::username().to_lowercase();
+        let clean_username: String = username.chars().filter(|c| c.is_alphanumeric()).collect();
+
+        println!("Cleaned username: {}", clean_username);
+
+        let protocol = config.protocol.to_string().to_lowercase();
+
+        let hashed_name = format!(
+            "kftray-forward-{}-{}-{}-{}",
+            clean_username, protocol, timestamp, random_string
+        )
+        .to_lowercase();
+
+        let config_id_str = config
+            .id
+            .map_or_else(|| "default".into(), |id| id.to_string());
+
+        if config
+            .remote_address
+            .as_ref()
+            .map_or(true, String::is_empty)
+        {
+            config.remote_address.clone_from(&config.service)
+        }
+
+        let mut values: HashMap<&str, String> = HashMap::new();
+        values.insert("hashed_name", hashed_name.clone());
+        values.insert("config_id", config_id_str);
+        values.insert("service_name", config.service.as_ref().unwrap().clone());
+        values.insert(
+            "remote_address",
+            config.remote_address.as_ref().unwrap().clone(),
+        );
+        values.insert("remote_port", config.remote_port.to_string());
+        values.insert("local_port", config.remote_port.to_string());
+        values.insert("protocol", protocol.clone());
+
+        let mut file = File::open(get_pod_manifest_path()).map_err(|e| e.to_string())?;
+        let mut contents = String::new();
+        file.read_to_string(&mut contents)
+            .map_err(|e| e.to_string())?;
+
+        let rendered_json = render_json_template(&contents, &values);
+        let pod: Pod = serde_json::from_str(&rendered_json).map_err(|e| e.to_string())?;
+
+        let pods: Api<Pod> = Api::namespaced(client.clone(), &config.namespace);
+
+        match pods.create(&kube::api::PostParams::default(), &pod).await {
+            Ok(_) => {
+                if let Err(e) = kube_runtime::wait::await_condition(
+                    pods.clone(),
+                    &hashed_name,
+                    conditions::is_pod_running(),
+                )
+                .await
+                {
+                    let dp = DeleteParams {
+                        grace_period_seconds: Some(0),
+                        ..DeleteParams::default()
+                    };
+                    let _ = pods.delete(&hashed_name, &dp).await;
+                    return Err(e.to_string());
+                }
+
+                config.service = Some(hashed_name.clone());
+
+                let start_response = match protocol.as_str() {
+                    "udp" => start_port_udp_forward(vec![config.clone()]).await,
+                    "tcp" => start_port_forward(vec![config.clone()]).await,
+                    _ => {
+                        let _ = pods
+                            .delete(&hashed_name, &kube::api::DeleteParams::default())
+                            .await;
+                        return Err("Unsupported proxy type".to_string());
+                    }
+                };
+
+                match start_response {
+                    Ok(mut port_forward_responses) => {
+                        let response = port_forward_responses
+                            .pop()
+                            .ok_or("No response received from port forwarding")?;
+                        responses.push(response);
+                    }
+                    Err(e) => {
+                        let _ = pods
+                            .delete(&hashed_name, &kube::api::DeleteParams::default())
+                            .await;
+                        return Err(format!("Failed to start port forwarding {}", e));
+                    }
+                }
+            }
+            Err(e) => return Err(e.to_string()),
+        }
+    }
+
+    Ok(responses)
+}
+
+#[tauri::command]
+
+pub async fn stop_proxy_forward(
+    config_id: String, namespace: &str, service_name: String,
+) -> Result<CustomResponse, String> {
+    log::info!(
+        "Attempting to stop proxy forward for service: {}",
+        service_name
+    );
+
+    let client = Client::try_default().await.map_err(|e| {
+        log::error!("Failed to create Kubernetes client: {}", e);
+
+        e.to_string()
+    })?;
+
+    let pods: Api<Pod> = Api::namespaced(client, namespace);
+
+    let lp = ListParams::default().labels(&format!("config_id={}", config_id));
+
+    let pod_list = pods.list(&lp).await.map_err(|e| {
+        log::error!("Error listing pods: {}", e);
+
+        e.to_string()
+    })?;
+
+    let username = whoami::username();
+
+    let pod_prefix = format!("kftray-forward-{}", username);
+
+    log::info!("Looking for pods with prefix: {}", pod_prefix);
+
+    for pod in pod_list.items {
+        if let Some(pod_name) = pod.metadata.name {
+            if pod_name.starts_with(&pod_prefix) {
+                log::info!("Found pod to stop: {}", pod_name);
+
+                let delete_options = DeleteParams {
+                    grace_period_seconds: Some(0),
+                    propagation_policy: Some(kube::api::PropagationPolicy::Background),
+                    ..Default::default()
+                };
+
+                match pods.delete(&pod_name, &delete_options).await {
+                    Ok(_) => log::info!("Successfully deleted pod: {}", pod_name),
+                    Err(e) => {
+                        log::error!("Failed to delete pod: {} with error: {}", pod_name, e);
+
+                        return Err(e.to_string());
+                    }
+                }
+
+                break;
+            } else {
+                log::info!("Pod {} does not match prefix, skipping", pod_name);
+            }
+        }
+    }
+
+    log::info!("Stopping port forward for service: {}", service_name);
+
+    let stop_result = stop_port_forward(service_name.clone(), config_id)
+        .await
+        .map_err(|e| {
+            log::error!(
+                "Failed to stop port forwarding for service '{}': {}",
+                service_name,
+                e
+            );
+
+            e
+        })?;
+
+    log::info!("Proxy forward stopped for service: {}", service_name);
+
+    Ok(stop_result)
+}
