+++ crates/kftray-tauri/src/kubeforward/mod.rs
@@ -1,9 +1,9 @@
-// kubeforward/mod.rs
-
+pub mod commands;
 pub mod kubecontext;
+pub mod logging;
 pub mod pod_selection;
 pub mod port_forward;
-pub mod proxy;
+pub mod target_pod_finder;
 
 use anyhow::Context;
 pub use k8s_openapi::api::core::v1 as vx;
