+++ crates/kftray-tauri/src/tray.rs
@@ -15,7 +15,7 @@ use tauri_plugin_positioner::Position;
 use tokio::runtime::Runtime;
 use tokio::time::sleep;
 
-use crate::kubeforward::port_forward;
+use crate::kubeforward::commands;
 use crate::models::window::SaveDialogState;
 use crate::window::{
     adjust_window_size_and_position,
@@ -179,7 +179,7 @@ pub fn stop_all_port_forwards_and_exit(app_handle: &tauri::AppHandle) {
     let runtime = Runtime::new().expect("Failed to create a Tokio runtime");
 
     runtime.block_on(async {
-        match port_forward::stop_all_port_forward().await {
+        match commands::stop_all_port_forward().await {
             Ok(_) => {
                 println!("Successfully stopped all port forwards.");
             }
