+++ crates/kftray-tauri/tauri.conf.json
